// src/app/features/products/enums/product-price-type.ts
export enum ProductPriceType {
  Base = 'BASE',               // قیمت پایه / اصلی
  Regional = 'REGIONAL',       // قیمت منطقه‌ای
  UserTier = 'USER_TIER',      // قیمت بر اساس نوع کاربر (عادی، VIP و...)
  Festival = 'FESTIVAL',       // قیمت جشنواره
  FlashSale = 'FLASH_SALE'     // قیمت شگفت‌انگیز
}